backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.11.0
pydantic-settings>=2.8.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
cryptography>=42.0.5
slowapi>=0.1.9
httpx>=0.27.0
openpyxl>=3.1.2
reportlab>=4.1.0
pandas>=2.2.1
python-dotenv>=1.0.1
email-validator>=2.1.1
aiosqlite>=0.20.0
pytz>=2024.1
google-genai>=1.0.0
pillow>=10.0.0
